feat(services)[travel-time-management]: add `GetTravelTimes` method
This method will get `loaderTypeId` and other things as optional. One of
these optional values should be pass to API.
It will return `TravelTime[]`.
The `isLoggedIn()` method added.
It will set undefined to not given value and it both goes undefined it
will set a 0 as sourse id value.
`LoaderTypeId` should be available, `SourceCityId` can be undefined,
`TargetCityId` can be undefined (one of both undefined should be
defiend)