feat(services)[tariffs-management]: add `GetTariffs` method
The method will get 4 parameters which will map to the `Tariff` model.
It will return array of `Tariff`.
After it checks the user session validation, It will handle the body of
our request.
The first parameter should be present. One of three others should be at
least have valid value. It will check every three parameter, and if it
goes okey, it will added to the body of our POST request. At the end of
checking, it will sent a default value.