feat(components/shared/button)[button.*]: add `Button` component
The new component, cutomed with linear background, animation, and etc.
It will support all aspects of p-button with more options.
styles added to an object to make the code format look like of using
`[dt]` option of `PrimeNg` components.
On init, it will make the style class base on options that it supports.
When the component goes disable, the click method and the animation stop
functioning.