feat(components/shared/main-view)[carousel.*]: add custom carousel
I used `p-carousel` from `PrimeNg`.
It will accept array of object for its images (specific model).
Its setting is a signal and can be handle dynamiclly in its component.ts
file.
For situations that the data goes less or equal to what it shows on, the
circular option goes false because circular would go on static images
and after clicking to the next one, it then shows the same images but
act like there is new ones.