fix(app/components/shared)[button]: update `styleHandler` method due to its bug
The style doesn't work on second time of disable change because the last
code only changed the style at first and for other times it added new
style instead of update the last style. This new method and signal will
solve the problem and will update the style instead of added new style
to it.