refactor(app/components/shared/)[support-button]: update style + add dynamic number list
Because we are going to use this panel in multiple provinces, so we need
a list of numbers which will specificly contains needed numbers of that
province. For now we are in Esfahan only but it should change in future
and has specific API controler that tells us what number we should show
in this support button.